+++ src/i18n/locales/en/common.json
@@ -103,7 +103,7 @@
 		"image_saved": "Image saved to {{path}}",
 		"mode_exported": "Mode '{{mode}}' exported successfully",
 		"mode_imported": "Mode imported successfully",
-		"command_whitelisted": "Command pattern '{{pattern}}' has been added to the allowed commands list",
+		"command_allowed": "Command pattern '{{pattern}}' has been added to the allowed commands list",
 		"command_denied": "Command pattern '{{pattern}}' has been added to the denied commands list"
 	},
 	"answers": {
