+++ src/i18n/locales/fr/common.json
@@ -103,7 +103,7 @@
 		"public_share_link_copied": "Lien de partage public copié dans le presse-papiers !",
 		"mode_exported": "Mode '{{mode}}' exporté avec succès",
 		"mode_imported": "Mode importé avec succès",
-		"command_whitelisted": "Le modèle de commande '{{pattern}}' a été ajouté à la liste des commandes autorisées",
+		"command_allowed": "Le modèle de commande '{{pattern}}' a été ajouté à la liste des commandes autorisées",
 		"command_denied": "Le modèle de commande '{{pattern}}' a été ajouté à la liste des commandes refusées"
 	},
 	"answers": {
