+++ src/i18n/locales/hi/common.json
@@ -103,7 +103,7 @@
 		"public_share_link_copied": "सार्वजनिक साझाकरण लिंक क्लिपबोर्ड में कॉपी किया गया!",
 		"mode_exported": "मोड '{{mode}}' सफलतापूर्वक निर्यात किया गया",
 		"mode_imported": "मोड सफलतापूर्वक आयात किया गया",
-		"command_whitelisted": "कमांड पैटर्न '{{pattern}}' को अनुमत कमांड सूची में जोड़ा गया है",
+		"command_allowed": "कमांड पैटर्न '{{pattern}}' को अनुमत कमांड सूची में जोड़ा गया है",
 		"command_denied": "कमांड पैटर्न '{{pattern}}' को अस्वीकृत कमांड सूची में जोड़ा गया है"
 	},
 	"answers": {
