+++ src/i18n/locales/it/common.json
@@ -103,7 +103,7 @@
 		"public_share_link_copied": "Link di condivisione pubblica copiato negli appunti!",
 		"mode_exported": "Modalità '{{mode}}' esportata con successo",
 		"mode_imported": "Modalità importata con successo",
-		"command_whitelisted": "Il modello di comando '{{pattern}}' è stato aggiunto all'elenco dei comandi consentiti",
+		"command_allowed": "Il modello di comando '{{pattern}}' è stato aggiunto all'elenco dei comandi consentiti",
 		"command_denied": "Il pattern di comando '{{pattern}}' è stato aggiunto all'elenco dei comandi negati"
 	},
 	"answers": {
