+++ src/i18n/locales/ja/common.json
@@ -103,7 +103,7 @@
 		"public_share_link_copied": "公開共有リンクがクリップボードにコピーされました！",
 		"mode_exported": "モード「{{mode}}」が正常にエクスポートされました",
 		"mode_imported": "モードが正常にインポートされました",
-		"command_whitelisted": "コマンドパターン '{{pattern}}' が許可されたコマンドリストに追加されました",
+		"command_allowed": "コマンドパターン '{{pattern}}' が許可されたコマンドリストに追加されました",
 		"command_denied": "コマンドパターン '{{pattern}}' が拒否されたコマンドリストに追加されました"
 	},
 	"answers": {
