+++ src/i18n/locales/ko/common.json
@@ -103,7 +103,7 @@
 		"public_share_link_copied": "공개 공유 링크가 클립보드에 복사되었습니다!",
 		"mode_exported": "'{{mode}}' 모드가 성공적으로 내보내졌습니다",
 		"mode_imported": "모드를 성공적으로 가져왔습니다",
-		"command_whitelisted": "명령 패턴 '{{pattern}}'이(가) 허용된 명령 목록에 추가되었습니다",
+		"command_allowed": "명령 패턴 '{{pattern}}'이(가) 허용된 명령 목록에 추가되었습니다",
 		"command_denied": "명령 패턴 '{{pattern}}'이(가) 거부된 명령 목록에 추가되었습니다"
 	},
 	"answers": {
