+++ src/i18n/locales/pl/common.json
@@ -103,7 +103,7 @@
 		"public_share_link_copied": "Publiczny link udostępniania skopiowany do schowka!",
 		"mode_exported": "Tryb '{{mode}}' pomyślnie wyeksportowany",
 		"mode_imported": "Tryb pomyślnie zaimportowany",
-		"command_whitelisted": "Wzór polecenia '{{pattern}}' został dodany do listy dozwolonych poleceń",
+		"command_allowed": "Wzór polecenia '{{pattern}}' został dodany do listy dozwolonych poleceń",
 		"command_denied": "Wzór polecenia '{{pattern}}' został dodany do listy odrzuconych poleceń"
 	},
 	"answers": {
