+++ src/i18n/locales/pt-BR/common.json
@@ -107,7 +107,7 @@
 		"public_share_link_copied": "Link de compartilhamento público copiado para a área de transferência!",
 		"mode_exported": "Modo '{{mode}}' exportado com sucesso",
 		"mode_imported": "Modo importado com sucesso",
-		"command_whitelisted": "O padrão de comando '{{pattern}}' foi adicionado à lista de comandos permitidos",
+		"command_allowed": "O padrão de comando '{{pattern}}' foi adicionado à lista de comandos permitidos",
 		"command_denied": "O padrão de comando '{{pattern}}' foi adicionado à lista de comandos negados"
 	},
 	"answers": {
