+++ src/i18n/locales/ru/common.json
@@ -103,7 +103,7 @@
 		"public_share_link_copied": "Публичная ссылка для совместного доступа скопирована в буфер обмена!",
 		"mode_exported": "Режим '{{mode}}' успешно экспортирован",
 		"mode_imported": "Режим успешно импортирован",
-		"command_whitelisted": "Шаблон команды '{{pattern}}' добавлен в список разрешенных команд",
+		"command_allowed": "Шаблон команды '{{pattern}}' добавлен в список разрешенных команд",
 		"command_denied": "Шаблон команды '{{pattern}}' добавлен в список запрещенных команд"
 	},
 	"answers": {
