+++ src/i18n/locales/zh-CN/common.json
@@ -108,7 +108,7 @@
 		"public_share_link_copied": "公开分享链接已复制到剪贴板！",
 		"mode_exported": "模式 '{{mode}}' 已成功导出",
 		"mode_imported": "模式已成功导入",
-		"command_whitelisted": "命令模式 '{{pattern}}' 已添加到允许的命令列表中",
+		"command_allowed": "命令模式 '{{pattern}}' 已添加到允许的命令列表中",
 		"command_denied": "命令模式 '{{pattern}}' 已被添加到拒绝命令列表中"
 	},
 	"answers": {
